Replaced `get` with `get_v`
Tested: OK
Before this commit, to access an element in the variables (inherited from map) I could use `map.at`, but for when later there is variable object you need to do `map::at().value()`.
Now the functions get does exactly this, when the object has the method value() it access it directly.