Copy constructor mustn't invalidate pointers
Copy constructor invalidates only pointers to object inside themselves (e.g., head pointing to the instance of head contained in the results attribute). If an attribute points to something, that relationship must stay there as the copy constructors are for passing to functions.

What we don't want is that the resulting objects from the cloning operations has pointers to other objects.
Example for a new node, it is fine that it still points to the links attached to when copied, as this function is called, for example, to create a copy of that object for another function or a const instance etc.
When you are creating a new node for another network, hence cloning the node, there is where you don't want to copy the links.