Tested clone for the curves
Build: Y
Run: Y
Test: TODO
Easy to clone the object. However, there was a bug in the copy constructor that was not copying the data object (which could actually become a property).
it works, but I was caching the indices after adding the element, and so it was crashing because I didn't copy the EPANET project.