Print anytown twophases problem runtime info
B[x] R[x] T[~]
I removed the templatization of the functions to write out the static description of the pagmo containers. They are known; I made all of them already, so there is no need to have a template there. It makes sense to have a template only for the UDCs because I have implemented only nsga2 and this specific problem.