Results behaviour changed
To represent the results (values.size() == timestep.size()+1), instead of allowing for having the last time step equal to the duratio and the values always equal to the size of the timestep vector, a QuantitySeries is now valid when it has either values.size() == timestep.size() or values.size() == timestep.size()+1. In the first case, when duration is accessed, you wrap around 0. While if values.size() == timestep.size()+1, when pos=length() or time__s=duration, you can access the last one because it exists.
B[x]
R[x]
T[~]