HOTFIX: install riser from tank to junction as before
At the beginning of the project I was installing the riser (the pipe connecting the tank with the junction) from the Tank to the Junction.
This is the same as doing it from the Junction to the Tank.
However, since EPANET initializes the flows with the direction provided by the user, I was ending up with slightly different results (circa 1e-8).
Reverted back and it works. This error will be investigated more in details later