Version 25.02.0
Updated EPANET to the last commit in development. This breaks the previous results and it will make you get different numbers from the simulations.
Therefore I updated the CMake to make the project work with the new version.
I also removed the possibility of changing EPANET build behaviour with environmental variables (i have never seen it in any of the projects).